cmd/tgfeed: initially refactor admin dashboard
