+++ cmd/tgfeed/internal/admin/frontend/components/ChartsGrid.tsx
@@ -1,0 +1,11 @@
+import React from "npm:react";
+
+export function ChartsGrid(props: {
+  children: React.ReactNode;
+}) {
+  return (
+    <div className="chart-matrix">
+      {props.children}
+    </div>
+  );
+}
